Hardcode the DConf for defaultuser.
oFono is run as root and GSettings is using
root's DConf entries to get the ignored and
blocked numbers.

This is not convenient for the user that
would like to choose which numbers to deal
with. So hard-code for the moment the
home directory of 'defaultuser'. This is
very very bad, but work for the moment.